Fix bug and remove unused language
We already have a branch for patching from the old repository.
This patch file (en-vi.patch) is from there. Because we lost
the local branch due to an accident removal of the repo, we
just use the old patch file without creating commits/branches.

I do now really know if this is a good style. At least it just
works