:import auto. checkout the package from multiple trees
In this version, the command (run.sh/import) will accept a list of ABS sources.
It will check every source if the package $_pkg is found. If it can not find
such package, it will run "svn update" in each source to check out the source

We also run "svn update" to get the latest package information from Arch
before importing the package to our repository