Fix dependencies issue that breaks the uim-setup
If the package m17n-db is installed at the time the package is compiled,
the package will insert all entries from m17n-db to loaded.scm. If the
user then removes the package m17n-db and uses the package m17n-db-vi,
the uim-setup still sees all entries from the old pkg. m17n-db. Be that
some files are missing, uim-setup will simplify terminate abnormally.

This issue is found by Quang Nguyen <luffy#archlinuxvn> while he is
building the package for his own 64-bit system.

The work-around is to fix the dependency issue as we are doing now.
The better way is to re-genereate uim files as needed. Those files are
 . /usr/share/uim/loaded.scm
 . /usr/share/uim/installed-modules.scm