Fix bug in _get_git_commits_between_two_points
The bug: If we use HEAD HEAD, there is actually no commit found;
however, bc the function still prints a "echo", the next func.
will see that there is (nil) commit, and the counter will see 1.

This bug is found when building a package whose the HEAD is on
the latest tag: the (s-makepkg) returns 2 in release-number (should 1)

To fix the bug, we will (grep) to get all valid tags. Bc (grep)
has no problem with EOF character, and (grep) even generates
an EOF for us, we will have a correct output. Please note that
bc. we do not use (silent) option for (grep), the grep return code
would be: 0 (line found) 1 (no line found) or 2 (grep error).
Both values 0 and 1 are valid in our case, bc. the purpose
is to use (grep) for output filtering, not to find something