In :source, replace :pkgname by :pkgbase
A reminder: :pkgname maybe an array, while :pkgbase reflects
exactly the :base of the package. Because :source is evaluated
very soon, the :pkgname may not be valid (TheSLinux only).
It took me > 30 minutes to figure out what :pkgname is in the
source tree of ArchLinux :makepkg.

If you have the same question remember to check the function
(run_split_packaging). In this function, an iterator named :pgkname
is used to go through the list of all sub-packages. Basically,
:pkgname has two different means:
  (1) a list of sub-packages
  (2) the name of package name that is being built.
That is why sometimes :pkgname is needed, but sometimes it is not.
I do not know why this is it, but I should do the same thing;
the variable is mutable, the variable may be changed, but if
we update its mean it may be terrible.

So much time for this stuff now I go for some drink ^^