Use :TheSmallBang as default
:TheBigBang is quite big, and we because it is big,
it may make users confused. TheSmallBang is really simple,
and its rules will never be out-of-date, and we do not
need to merge/rebase. TheSmallBang is freeze, it should never
be updated in any situation, while :TheBigBang is developed
and it is used to provide the latest development guides.

I do not why people call it the :Big. The space is too small,
it can not be big. Let us give it a new name :)