Fix :get_update to use new Github format
The output format of Github page /tags/ has changed (!?),
and we need to use another :regexp to filter real tags.

This is an example of how hard the :get_update can be.
It is also clear that though we can use the old PKGBUILD,
the portion :get_update should only be in the latest branch.
Maybe we need to move :get_update off PKGBUILD