security fix: :fcrontab must not run as :root
The "--with-{user,group}name=" flags specify the user under that
the ":fcrontab" program runs. This program should not run as :root;
otherwise, any trivial user on the system can use the ":runas" field
to escape their own privilege, and to be :root.

These flags should never be ":root" or ":wheel" group.

Please note that ":fcrontab" is a :suid program. On :cronie, :crontab
belongs to root, has :suid flag, and it just works without any problem.
I think the design of :fcron is different.

But it should not work like this.

Please note this :pkgbuild is not complete. Other things to do:
* Add new user (:fcron)
* Fix some file permissions (/etc/fcron/fcron.conf -> root:fcron)
* Fix dir. and files permissions in /var/spool/fcron