Remove the use of $pkgdir to avoid some warning
In (makepkg::tidy_install), there is a check if $srcdir is
referred somewhere in the $pkgdir. However, this check is
done on all files in $pkgdir, that is not very useful. APR
package, for example, use $srcdir in its :configure script,
and that will generate some alerts after package is built.

By removing some use of $pkgdir in PKGBUILD, we make sure
that the old style is not applied, but the warning is still
there there is not an easy to avoid