Allow :modifier support in PACKAGE_FEATURE
PACKAGE_FEATURE now *should* be started by a modifier, which
is @, #, =, % or +. The modifier gives information about the
conflict status of {package, feature package}. Possible values:

  @: conflict
  %: not conflict
  =: noop (developer will handle this manually)

The variable PACKAGE_CONFLICT_TYPE will be set when :_s_env
is executed. In :s-makepkg, this variable must be handled.

Primary purpose: To build a sub-package, or part of the package