!bug: can't build #uim-vi-1.8.6
There are some differences between #git source tree
and the tarball distributed on #google hosting. E.g,

* ./scm/trec-composer.scm found on #git, not on #google
* ./pixmaps/viqr.svg      found on #google, not on #git

In this PKGBUILD, we will use the tarball from #github,
and we also use the tarball from #uim-1.8.6 distribution
to fetch some pixmap files and #sigscheme library. The
pixmap files are stored some secret place :)

This PKGBUILD is now working with both #uim-vi-1.8.{5,6}.