pkgbuild#_pkgbuild_load: quick support for #lib32 packages
This commit provides an ability to compile #lib32 package on 64-bit
machine. This is not to build 32-bit package on 64-bit machine.

Example, #libdatrie%lib32 provides basic libraries for 32-bit apps.
to run on 64-bit machines. The final package will add #conflict
and #provide information automatically. So that,

  #libdatrie%lib32-test conflicts with #libdatrie%lib32

Please note that, %lib32 is the correct feature. We do not use
@lib32 here, because the #libdatrie and #libdatrie%lib32 coexists.