pkgbuild#_pkgbuild_load: fix $CXX for %lib32 support
For %lib32, we manually add "-m32" flag to $CXX. This is because
some linking program may not read "-m32" from $CFLAGS. For example,
in #uim, by setting "QMAKE_LINK = @CXX@" in "common.pro.in", the
#qmake program will tell #qt to use "g++" for linking. The correct
linker should be "g++ -m32".

FIXME: I am not sure this is the best/correct way.