lib/git#_get_package_name_from_tag: return #commits if rel. number is zero
When zero is used as release number in PACKAGE_TAG, e,g, x.y.z-0, the method
will detect number of commits since TheSmallBang. This is useful when you are
providing a *future* tag to #s-makepkg. For example, if you are working on 1.9
branch, and a future release is 1.9.0, you can use PACKAGE_TAG=1.9-0; #s-makepkg
will translate to final package version 1.9-59 where 59 is the number of commits
since TheSmallBang.

Please note, if -0 is not provide, e.g, PACKAGE_TAG=x.y, release number is 1.